requirements.txt
transformers>=4.30.0
optimum>=1.7.0
onnx>=1.13.0
onnxruntime-directml>=1.12.0
GPUtil>=1.4.0
better_profanity>=0.7.0
sacremoses>=0.0.53
gitpython>=3.1.31
torch>=1.12.0
requests>=2.28.0
srt>=1.5.1
